Removing println comments and starting cleanup
